# src/main/kotlin/com/example/dobby/controller/web/DashboardController.kt
package com.example.dobby.controller.web

import com.example.dobby.dto.ApiResponse
import com.example.dobby.dto.roast.RoastLogDbResponse
import com.example.dobby.dto.user.UserProfileResponse
import com.example.dobby.service.DashboardService
import org.springframework.web.bind.annotation.*

@RestController
@RequestMapping("/dashboard")
class DashboardController(
    private val dashboardService: DashboardService
) {

    @GetMapping("facts/{guildId}")
    suspend fun getAllGuildFacts(
        @PathVariable("guildId") guildId: String
    ): List<UserProfileResponse> {
        return dashboardService.getAllFactsByGuildId(guildId)
    }

    @DeleteMapping("facts/{factId}")
    suspend fun deleteGuildFact(
        @PathVariable("factId") factId: String
    ): ApiResponse {
        dashboardService.deleteFact(factId)
        return ApiResponse(
            success = true,
            message = "Fact deleted successfully."
        )
    }

    @DeleteMapping("facts/{guildId}/reset")
    suspend fun resetGuildFacts(
        @PathVariable("guildId") guildId: String
    ): ApiResponse {
        dashboardService.purgeGuildFacts(guildId)
        return ApiResponse(
            success = true,
            message = "Fact deleted successfully."
        )
    }


    @GetMapping("roasts/{guildId}")
    suspend fun getAllGuildRoasts(
        @PathVariable("guildId") guildId: String
    ): List<RoastLogDbResponse> {
        return dashboardService.getAllRoastByGuildId(guildId)
    }

}