# src/main/kotlin/com/example/dobby/dto/user/UserProfile.kt
package com.example.dobby.dto.user

import com.example.dobby.dto.fact.UserFactResponse
import kotlinx.serialization.SerialName
import kotlinx.serialization.Serializable


@Serializable
data class UserProfileCreateRequest(
    @SerialName("discord_user_id")
    val discordUserId: String,
    @SerialName("guild_id")
    val guildId: String,
    @SerialName("display_name")
    val displayName: String?,
    @SerialName("avatar_hash")
    val avatarHash: String?
)


@Serializable
data class UserProfileResponse(
    @SerialName("id")
    val id: String,
    @SerialName("discord_user_id")
    val discordUserId: String,
    @SerialName("guild_id")
    val guildId: String,
    @SerialName("display_name")
    val displayName: String?,
    @SerialName("avatar_hash")
    val avatarHash: String?,
    @SerialName("created_at")
    val createdAt: String,
    @SerialName("updated_at")
    val updatedAt: String?,
    @SerialName("user_facts")
    val facts: List<UserFactResponse>? = emptyList()
)